Bug 777614 - Update Skia to r5539.
UltraBlame original commit: 81b2a9d689799669d2635daff3ede4fccdd512a6

+++ gfx/skia/src/images/SkFDStream.cpp
@@ -57,7 +57,7 @@ size_t SkFDStream::read(void* buffer, size_t size) {
 #endif
                 return 0;
             }
-            return size;
+            return (size_t) size;
         } else if (NULL == buffer) {        
             off_t oldCurr = ::lseek(fFD, 0, SEEK_CUR);
             if (oldCurr < 0) {
@@ -74,7 +74,7 @@ size_t SkFDStream::read(void* buffer, size_t size) {
                 return 0;   
             }
             
-            return newCurr - oldCurr;
+            return (size_t) (newCurr - oldCurr);
         } else {                            
             ssize_t actual = ::read(fFD, buffer, size);
